commit gfx api commits
