+++ cocos/gfx/binding-layout.ts
@@ -4,16 +4,28 @@ import { GFXDevice } from './device';
 import { GFXSampler } from './sampler';
 import { GFXTextureView } from './texture-view';
 
+/**
+ * @zh
+ * GFX绑定
+ */
 export interface IGFXBinding {
     binding: number;
     type: GFXBindingType;
     name: string;
 }
 
+/**
+ * @zh
+ * GFX绑定布局描述信息
+ */
 export interface IGFXBindingLayoutInfo {
     bindings: IGFXBinding[];
 }
 
+/**
+ * @zh
+ * GFX绑定单元
+ */
 export class GFXBindingUnit {
     public binding: number = 0;
     public type: GFXBindingType = GFXBindingType.UNKNOWN;
@@ -23,21 +35,65 @@ export class GFXBindingUnit {
     public sampler: GFXSampler | null = null;
 }
 
+/**
+ * @zh
+ * GFX绑定布局
+ */
 export abstract class GFXBindingLayout extends GFXObject {
 
+    /**
+     * @zh
+     * GFX设备
+     */
     protected _device: GFXDevice;
+
+    /**
+     * @zh
+     * 绑定单元数组
+     */
     protected _bindingUnits: GFXBindingUnit[] = [];
+
+    /**
+     * @zh
+     * 脏数据标识
+     */
     protected _isDirty = false;
 
+    /**
+     * @zh
+     * 构造函数
+     * @param device GFX设备
+     */
     constructor (device: GFXDevice) {
         super(GFXObjectType.BINDING_LAYOUT);
         this._device = device;
     }
 
+    /**
+     * @zh
+     * 初始化函数
+     * @param info GFX绑定布局描述信息
+     */
     public abstract initialize (info: IGFXBindingLayoutInfo): boolean;
+
+    /**
+     * @zh
+     * 销毁函数
+     */
     public abstract destroy ();
+
+    /**
+     * @zh
+     * 更新
+     */
     public abstract update ();
 
+    /**
+     * @zh
+     * 在指定的binding位置上绑定缓冲
+     * @param binding 绑定GFX组件的插槽
+     * @param buffer GFX缓冲
+     */
     public bindBuffer (binding: number, buffer: GFXBuffer) {
         for (const bindingUnit of this._bindingUnits) {
             if (bindingUnit.binding === binding) {
@@ -54,6 +110,12 @@ export abstract class GFXBindingLayout extends GFXObject {
         }
     }
 
+    /**
+     * @zh
+     * 在指定的binding位置上绑定采样器
+     * @param binding 绑定GFX组件的插槽
+     * @param sampler GFX采样器
+     */
     public bindSampler (binding: number, sampler: GFXSampler) {
         for (const bindingUnit of this._bindingUnits) {
             if (bindingUnit.binding === binding) {
@@ -70,6 +132,12 @@ export abstract class GFXBindingLayout extends GFXObject {
         }
     }
 
+    /**
+     * @zh
+     * 在指定的binding位置上绑定纹理视图
+     * @param binding 绑定GFX组件的插槽
+     * @param texView GFX纹理视图
+     */
     public bindTextureView (binding: number, texView: GFXTextureView) {
         for (const bindingUnit of this._bindingUnits) {
             if (bindingUnit.binding === binding) {
@@ -86,6 +154,11 @@ export abstract class GFXBindingLayout extends GFXObject {
         }
     }
 
+    /**
+     * @zh
+     * 得到指定的binding位置上的GFX绑定单元
+     * @param binding 绑定GFX组件的插槽
+     */
     public getBindingUnit (binding: number): GFXBindingUnit | null {
         for (const unit of this._bindingUnits) {
             if (unit.binding === binding) {
