+++ cocos/gfx/render-pass.ts
@@ -9,6 +9,10 @@ import {
 } from './define';
 import { GFXDevice } from './device';
 
+/**
+ * @zh
+ * GFX颜色附件
+ */
 export class GFXColorAttachment {
     public format: GFXFormat = GFXFormat.UNKNOWN;
     public loadOp: GFXLoadOp = GFXLoadOp.CLEAR;
@@ -18,6 +22,10 @@ export class GFXColorAttachment {
     public endLayout: GFXTextureLayout = GFXTextureLayout.COLOR_ATTACHMENT_OPTIMAL;
 }
 
+/**
+ * @zh
+ * GFX深度模板附件
+ */
 export class GFXDepthStencilAttachment {
     public format: GFXFormat = GFXFormat.UNKNOWN;
     public depthLoadOp: GFXLoadOp = GFXLoadOp.CLEAR;
@@ -29,6 +37,10 @@ export class GFXDepthStencilAttachment {
     public endLayout: GFXTextureLayout = GFXTextureLayout.DEPTH_STENCIL_ATTACHMENT_OPTIMAL;
 }
 
+/**
+ * @zh
+ * GFX子过程描述信息
+ */
 export interface IGFXSubPassInfo {
     bindPoint: GFXPipelineBindPoint;
     inputs: number[];
@@ -38,24 +50,61 @@ export interface IGFXSubPassInfo {
     preserves: number[];
 }
 
+/**
+ * @zh
+ * GFX渲染过程描述信息
+ */
 export interface IGFXRenderPassInfo {
     colorAttachments?: GFXColorAttachment[];
     depthStencilAttachment?: GFXDepthStencilAttachment;
     // subPasses? : GFXSubPassInfo[];
 }
 
+/**
+ * @zh
+ * GFX渲染过程
+ */
 export abstract class GFXRenderPass extends GFXObject {
 
+    /**
+     * @zh
+     * GFX设备
+     */
     protected _device: GFXDevice;
+
+    /**
+     * @zh
+     * GFX颜色附件数组
+     */
     protected _colorInfos: GFXColorAttachment[] = [];
+
+    /**
+     * @zh
+     * GFX深度模板附件
+     */
     protected _depthStencilInfo: GFXDepthStencilAttachment | null = null;
 
+    /**
+     * @zh
+     * 构造函数
+     * @param device GFX设备
+     */
     constructor (device: GFXDevice) {
         super(GFXObjectType.RENDER_PASS);
         this._device = device;
     }
 
+    /**
+     * @zh
+     * 提交命令缓冲数组
+     * @param info GFX渲染过程描述信息
+     */
     public abstract initialize (info: IGFXRenderPassInfo): boolean;
+
+    /**
+     * @zh
+     * 销毁函数
+     */
     public abstract destroy (): void;
     // protected _subPasses : GFXSubPassInfo[] = [];
 }
