+++ cocos/gfx/sampler.ts
@@ -1,6 +1,10 @@
 import { GFXAddress, GFXComparisonFunc, GFXFilter, GFXObject, GFXObjectType, IGFXColor } from './define';
 import { GFXDevice } from './device';
 
+/**
+ * @zh
+ * GFX采样器描述信息
+ */
 export interface IGFXSamplerInfo {
     name?: string;
     minFilter?: GFXFilter;
@@ -17,6 +21,10 @@ export interface IGFXSamplerInfo {
     mipLODBias?: number;
 }
 
+/**
+ * @zh
+ * GFX采样器状态
+ */
 export class GFXSamplerState {
     public name: string = '';
     public minFilter: GFXFilter = GFXFilter.LINEAR;
@@ -32,6 +40,11 @@ export class GFXSamplerState {
     public maxLOD: number = 0;
     public mipLODBias: number = 0.0;
 
+    /**
+     * @zh
+     * 比较函数
+     * @param state GFX采样器状态
+     */
     public compare (state: GFXSamplerState): boolean {
         return (this.minFilter === state.minFilter) &&
         (this.magFilter === state.magFilter) &&
@@ -51,20 +64,51 @@ export class GFXSamplerState {
     }
 }
 
+/**
+ * @zh
+ * GFX采样器
+ */
 export abstract class GFXSampler extends GFXObject {
 
+    /**
+     * @zh
+     * GFX采样器状态
+     */
     public get state (): GFXSamplerState {
         return this._state;
     }
 
+    /**
+     * @zh
+     * GFX设备
+     */
     protected _device: GFXDevice;
+
+    /**
+     * @zh
+     * GFX采样器状态
+     */
     protected _state: GFXSamplerState = new GFXSamplerState();
 
+    /**
+     * @zh
+     * 构造函数
+     */
     constructor (device: GFXDevice) {
         super(GFXObjectType.SAMPLER);
         this._device = device;
     }
 
+    /**
+     * @zh
+     * 初始化函数
+     * @param info GFX采样器描述信息
+     */
     public abstract initialize (info: IGFXSamplerInfo): boolean;
+
+    /**
+     * @zh
+     * 销毁函数
+     */
     public abstract destroy (): void;
 }
